requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
apscheduler==3.10.4
gunicorn==21.2.0
requests==2.31.0

# Authentication
werkzeug>=3.0.0

# Rate Limiting
flask-limiter==3.5.0

# Validation
email-validator==2.1.0

# CSRF Protection
flask-wtf==1.2.1

# Ghost CMS Export
python-slugify==8.0.1

# oEmbed Image Processing
Pillow==10.2.0

# CORS Support
flask-cors==4.0.0
